fix(form): move shared components into shared folder
