+++ packages/paste-core/components/form/src/FormInput.tsx
@@ -2,81 +2,9 @@ import * as React from 'react';
 import * as PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import css from '@styled-system/css';
-import {Box} from '@twilio-paste/box';
-import {BoxShadow} from '@twilio-paste/style-props';
-
-interface InputWrapperProps {
-  disabled?: boolean;
-  hasError?: boolean;
-  readOnly?: boolean;
-}
-
-const InputWrapper: React.FC<InputWrapperProps> = ({disabled, hasError, readOnly, children}) => {
-  let boxShadow = 'shadowBorderInput';
-  if (disabled) {
-    boxShadow = 'shadowBorderInputDisabled';
-  } else if (hasError) {
-    boxShadow = 'shadowBorderInputError';
-  }
-
-  return (
-    <Box
-      display="flex"
-      width="100%"
-      backgroundColor={readOnly || disabled ? 'colorBackground' : 'colorBackgroundBody'}
-      boxShadow={boxShadow as BoxShadow}
-      borderRadius="borderRadius20"
-      transition="box-shadow 100ms ease-in"
-      cursor={disabled ? 'not-allowed' : 'text'}
-      _hover={{
-        boxShadow: hasError ? 'shadowBorderInputErrorHover' : 'shadowBorderInputHover',
-      }}
-      _focusWithin={{
-        boxShadow: 'shadowFocus',
-      }}
-    >
-      {children}
-    </Box>
-  );
-};
-
-const Prefix: React.FC = ({children}) => {
-  if (children == null) return null;
-  return (
-    <Box
-      display="flex"
-      alignItems="center"
-      backgroundColor="colorBackground"
-      padding="space30"
-      borderRightWidth="borderWidth10"
-      borderRightColor="colorBorderLight"
-      borderRightStyle="solid"
-      borderTopLeftRadius="borderRadius20"
-      borderBottomLeftRadius="borderRadius20"
-    >
-      {children}
-    </Box>
-  );
-};
-
-const Suffix: React.FC = ({children}) => {
-  if (children == null) return null;
-  return (
-    <Box
-      display="flex"
-      alignItems="center"
-      backgroundColor="colorBackground"
-      padding="space30"
-      borderLeftWidth="borderWidth10"
-      borderLeftColor="colorBorderLight"
-      borderLeftStyle="solid"
-      borderTopRightRadius="borderRadius20"
-      borderBottomRightRadius="borderRadius20"
-    >
-      {children}
-    </Box>
-  );
-};
+import {FieldWrapper} from './shared/FieldWrapper';
+import {Prefix} from './shared/Prefix';
+import {Suffix} from './shared/Suffix';
 
 export type FormInputTypes = 'text' | 'email' | 'hidden' | 'number' | 'password' | 'search' | 'tel';
 
@@ -153,7 +81,7 @@ const FormInput = React.forwardRef<HTMLInputElement, FormInputProps>(
     }
 
     return (
-      <InputWrapper readOnly={readOnly} disabled={disabled} hasError={hasError}>
+      <FieldWrapper readOnly={readOnly} disabled={disabled} hasError={hasError}>
         <Prefix>{insertBefore}</Prefix>
         <InputElement
           aria-invalid={hasError}
@@ -169,7 +97,7 @@ const FormInput = React.forwardRef<HTMLInputElement, FormInputProps>(
           readOnly={readOnly}
         />
         <Suffix>{insertAfter}</Suffix>
-      </InputWrapper>
+      </FieldWrapper>
     );
   }
 );
